add getUsers functionality - with refactoring
