+++ _controllers/articles.controllers.js
@@ -1,4 +1,4 @@
-const { selectArticleById, selectUsers } = require("../_models/app.models");
+const { selectArticleById } = require("../_models/articles.models");
 
 exports.getArticleById = (req, res, next) => {
   const articleId = req.params.article_id;
@@ -9,16 +9,6 @@ exports.getArticleById = (req, res, next) => {
       } else {
         res.status(200).send(article);
       }
-    })
-    .catch((err) => {
-      next(err);
-    });
-};
-
-exports.getUsers = (req, res) => {
-  selectUsers()
-    .then((users) => {
-      res.status(200).send(users);
     })
     .catch((err) => {
       next(err);
