+++ _models/users.controllers.js
@@ -1,0 +1,7 @@
+const db = require("../db/connection");
+
+exports.selectUsers = () => {
+  return db.query("SELECT * FROM users;").then((users) => {
+    return users.rows;
+  });
+};
